refactor: extract pool management to worker_pool.go
